+++ test/integration/integration_test.cc
@@ -949,21 +949,14 @@ TEST_P(IntegrationTest, ProcessObjectUnealthy) {
 
 TEST_P(IntegrationTest, TrailersDroppedDuringEncoding) { testTrailers(10, 10, false, false); }
 
-TEST_P(IntegrationTest, TrailersEncodedUpstream) {
-  // Enable the encoding of trailers upstream
-  config_helper_.addConfigModifier([&](envoy::config::bootstrap::v2::Bootstrap& bootstrap) -> void {
-    RELEASE_ASSERT(bootstrap.mutable_static_resources()->clusters_size() == 1, "");
-    if (fake_upstreams_[0]->httpType() == FakeHttpConnection::Type::HTTP1) {
-      auto* cluster = bootstrap.mutable_static_resources()->mutable_clusters(0);
-      cluster->mutable_http_protocol_options()->set_enable_trailers(true);
-    }
-  });
-  testTrailers(10, 10, true, false);
+TEST_P(IntegrationTest, TrailersDroppedUpstream) {
+  config_helper_.addConfigModifier(setEnableDownstreamTrailersHttp1());
+  testTrailers(10, 10, false, false);
 }
 
-TEST_P(IntegrationTest, TrailersEncodedDownstream) {
-  config_helper_.addConfigModifier(setEnableEncodeTrailersHttp1());
-  testTrailers(10, 10, false, true);
+TEST_P(IntegrationTest, TrailersDroppedDownstream) {
+  config_helper_.addConfigModifier(setEnableUpstreamTrailersHttp1());
+  testTrailers(10, 10, false, false);
 }
 
 INSTANTIATE_TEST_SUITE_P(IpVersions, UpstreamEndpointIntegrationTest,